requirements.txt
django
djangorestframework
python-decouple
opencv-python
numpy
cv2
pylint